% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{breastTCGA_ER}
\alias{breastTCGA_ER}
\title{ER positive or negative status of breast tumor samples}
\format{A Factor of length 887 of levels Negative and Positive (with 111 missing values for the normals).}
\source{
\url{https://cancergenome.nih.gov/}
}
\usage{
breastTCGA_ER
}
\description{
A factor indicating whether 887 breast samples in breastTCGA_Mat are ER positive or ER negative.
The matched normals have empty values.
}
\keyword{datasets}
