% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{breastTCGA_Group}
\alias{breastTCGA_Group}
\title{Normal or Tumor status of breast samples}
\format{A Factor of length 887 of levels NORMAL and TUMOR.}
\source{
\url{https://cancergenome.nih.gov/}
}
\usage{
breastTCGA_Group
}
\description{
A factor indicating whether 887 breast samples in breastTCGA_Mat are tumor or matched normal.
}
\keyword{datasets}
