% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{breastTCGA_Mat}
\alias{breastTCGA_Mat}
\title{Gene expression for 1000 genes in 887 breast samples}
\format{A data matrix with 1000 rows and 887 columns.}
\source{
\url{https://cancergenome.nih.gov/}
}
\usage{
breastTCGA_Mat
}
\description{
A data matrix containing a subset of the TCGA breast cancer dataset, with the gene level expression estimates 
in log2 transcripts per million for 887 breast samples.
}
\keyword{datasets}
