% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{computeChiSquaredTest}
\alias{computeChiSquaredTest}
\title{Compute chi-squared test}
\usage{
computeChiSquaredTest(Mat, Groups, classes)
}
\arguments{
\item{Mat}{Matrix of digitized binary or ternary data with each column corresponding to a sample and each 
row corresponding to a feature}

\item{Groups}{Factor indicating class association of samples}

\item{classes}{Vector of class labels; the test will be applied between the classes given.}
}
\description{
Given a binary or ternary data matrix with class associations of samples, computes chi-squared tests
for each feature between given groups
}
\examples{
data(divergence)
# select half of normals for baseline computation

sel = sample(which(breastTCGA_Group=="NORMAL"), floor(sum(breastTCGA_Group=="NORMAL") * 0.5))

D = computeDigitization(Mat=breastTCGA_Mat[, -sel], baseMat=breastTCGA_Mat[, sel])

C = computeChiSquaredTest(Mat=D$Mat.div, Groups=breastTCGA_Group[sel], classes=c("NORMAL", "TUMOR"))

}
\keyword{chi-squared}
