% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/declareImports.R
\name{declareImports}
\alias{declareImports}
\title{This R file contains Roxygen2 directives for importing functions from the highfrequency library}
\description{
This R file contains Roxygen2 directives for importing functions from the highfrequency library
}

