% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/kernelFunctions.R
\name{deltaKernels}
\alias{deltaKernels}
\alias{expDoubleKernel}
\title{Kernel functions for delta sequences}
\usage{
expDoubleKernel(x0, cnv.data, delta.bandwidth)
}
\arguments{
\item{x0}{kernel evaluation argument}

\item{cnv.data}{data for kernel-based convolution}

\item{delta.bandwidth}{kernel bandwiwdth}

\item{normalize}{if \code{TRUE}, the resulting weights will be normalized so that \code{sum(expDoubleKernel(x0,cnv.data,0.01)*0.001)}}
}
\description{
Functions for constructing delta-sequences, based on kernels, for spot estimation from HF data.
}

