% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{msigdb_Hallmarks}
\alias{msigdb_Hallmarks}
\title{Cancer Hallmark gene sets from the MSigDB collection}
\format{A list of length 17, with the hallmark gene set name, each a character vector of gene symbols.}
\source{
\url{https://http://software.broadinstitute.org/gsea/msigdb/}
}
\usage{
msigdb_Hallmarks
}
\description{
A subset of the cancer hallmarks functional gene sets from the MSigDB collection.
}
\keyword{datasets}
