% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/multipowerVariation.R
\name{multipowerVariation}
\alias{multipowerVariation}
\alias{rMPV}
\title{Multipower variation}
\usage{
rMPV(rdata, r.tot, num.int, makeReturns, year.days, align.by, align.period, ...)
}
\arguments{
\item{rdata}{an \code{xts} object containing 1 return series.}

\item{r.tot}{which power-variation to estimate.}

\item{num.int}{how many neighbouring intervals are taken into consideration.}

\item{makeReturns}{boolean, should be \code{TRUE} when price data is supplied. Defaults to \code{FALSE}.}

\item{align.by}{}

\item{align.period}{}

\item{...}{Arguments passed on to \code{\link[highfrequency]{aggregatePrice}}}
}
\description{
Calculate multipower variation, an estimator of integrated volatiltiy. See summary in Veraart (2010) and Jacod (2006).
}
\details{
The most important arguments to pass go \code{\link[highfrequency]{aggregatePrice}} are \code{marketopen} and \code{marketclose}, see documentation therein. The default values are different from our test data set. \cr \cr
Choose \code{r.tot} and \code{num.int} such that \code{r.tot/num.int < 2}.
}

