% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotInference.R
\name{plotInference}
\alias{plot.divinf}
\alias{plotInference}
\title{Plot realized divergence inference results}
\usage{
\method{plot}{divinf}(x, true = NULL, quantile = 0.95, ...)
}
\arguments{
\item{x}{list with fields \code{rDiv} and \code{asy.var}.}

\item{true}{xts time series.}

\item{quantile}{for confidence bounds}

\item{...}{graphical parameters}
}
\description{
This replaces the plot interface for objects of class \code{divinf}, output from the divergence estimating functions with asymptotic variance estimates.
}

