% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/realizedDivergence.R
\name{realizedDivergence}
\alias{rDiv}
\alias{realizedDivergence}
\title{Realized Divergence}
\usage{
rDiv(rdata, pow, makeReturns, align.by, align.period, intradaySeasonFun, ...)
}
\arguments{
\item{rdata}{an \code{xts} object containing 1 return series.}

\item{pow}{\code{numeric} vector of length \code{P}: divergences will be calculated at all powers.}

\item{makeReturns}{boolean, should be \code{TRUE} when price data is supplied. Defaults to \code{FALSE}.}

\item{align.by}{}

\item{align.period}{}

\item{...}{Arguments passed on to \code{\link[highfrequency]{aggregatePrice}}}
}
\description{
Functions for calculating realized weighted univariate divergences with power \code{pow}, and the variance of the estimates.
}
\details{
The most important arguments to pass go \code{\link[highfrequency]{aggregatePrice}} are \code{marketopen} and \code{marketclose}, see documentation therein. The default values are different from our test data set. \cr
\cr
\code{rDivInference} calculates the estimator and asymptotic variance based on the feasible estimator in Khajavi, Or{\l}owski and Trojani (2015). \cr \cr
\code{rDivTrueInference} calculates the estimator and the infeasible asymptotic variance based on jump data from simulation.
}

