% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/realizedJumpKurt.R
\name{realizedJumpKurt}
\alias{rJKurt}
\alias{realizedJumpKurt}
\title{Realized jump kurtosis divergence}
\usage{
rJKurt(rdata, pow, makeReturns, align.by, align.period, intradaySeasonFun, ...)
}
\arguments{
\item{rdata}{an \code{xts} object containing 1 return series.}

\item{pow}{\code{numeric} vector of length \code{P}: jump skew divergences will be calculated around all powers.}

\item{makeReturns}{boolean, should be \code{TRUE} when price data is supplied. Defaults to \code{FALSE}.}

\item{align.by}{}

\item{align.period}{}

\item{...}{Arguments passed on to \code{\link[highfrequency]{aggregatePrice}}}
}
\description{
Functions for calculating realized weighted univariate jump kurtosis divergences around divergence with power \code{pow}, and the variance of the estimates.
}
\details{
The most important arguments to pass go \code{\link[highfrequency]{aggregatePrice}} are \code{marketopen} and \code{marketclose}, see documentation therein. The default values are different from our test data set. \cr \cr
\code{rJKurt} calculates the realized kurtosis jump divergence. \cr \cr
\code{rJKurtInference} calls \code{rJKurt} and calculates the asymptotic variance of the estimates with the use of the feasible estimator. \cr \cr
\code{rJKurtTrueInference} calculates the realized measure and true asymptotic variance from simulated data, including jump data.
}

