% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{computeMultivariateBinaryMatrix}
\alias{computeMultivariateBinaryMatrix}
\title{Compute the binary matrix with digitized divergence coding}
\usage{
computeMultivariateBinaryMatrix(Mat, Baseline)
}
\arguments{
\item{Mat}{Matrix of data to be digitized, in [0, 1], with each column corresponding to a sample and each 
row corresponding to a feature; usually in quantile form.}

\item{Baseline}{A Baseline object; this corresponds to the output of findMultivariateGammaWithSupport() or 
computeMultivariateSupport()}
}
\value{
A matrix with the same columns as Mat, with rows being the multivariate features, containing the binary form data.
}
\description{
Function for obtaining the binary form for a matrix for multivariate divergence of data given a baseline range
}
\examples{
baseMat = breastTCGA_Mat[, breastTCGA_Group == "NORMAL"]
baseMat.q = computeQuantileMatrix(baseMat)
baseline = computeMultivariateSupport(Mat=baseMat.q, FeatureSets=msigdb_Hallmarks)
dataMat = breastTCGA_Mat[, breastTCGA_Group != "NORMAL"]
dataMat.q = computeQuantileMatrix(dataMat)
divMat = computeMultivariateBinaryMatrix(Mat=dataMat.q, Baseline=baseline)

}
\keyword{binary}
\keyword{digitization}
