% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{computeUnivariateTernaryMatrix}
\alias{computeUnivariateTernaryMatrix}
\title{Compute the ternary matrix with digitized divergence coding}
\usage{
computeUnivariateTernaryMatrix(Mat, Baseline)
}
\arguments{
\item{Mat}{Matrix of data to be digitized, in [0, 1], with each column corresponding to a sample and each 
row corresponding to a feature; usually in quantile form.}

\item{Baseline}{A list with a data frame element "Ranges" containing the baseline range of each features; 
this corresponds to the output of findUnivariateGammaWithSupport() or computeUnivariateSupport()}
}
\value{
A matrix with the same dimensions as Mat containing the ternary form data.
}
\description{
Function for obtaining the ternary form for a matrix of data given a baseline range
}
\examples{
baseMat = breastTCGA_Mat[, breastTCGA_Group == "NORMAL"]
baseMat.q = computeQuantileMatrix(baseMat)
baseline = computeUnivariateSupport(Mat=baseMat.q)
dataMat = breastTCGA_Mat[, breastTCGA_Group != "NORMAL"]
dataMat.q = computeQuantileMatrix(dataMat)
divMat = computeUnivariateTernaryMatrix(Mat=dataMat.q, Baseline=baseline)

}
\keyword{digitization}
\keyword{ternary}
